Model Classes created
I created all the model classes that represent the long term object of
our game. The game I am creating first will only use one card but the
app is designed in a generic way in order to facilitate the
implementation of future card games with a more complex business logic
involved.

But first let's keep it simple!